vespers - last antiphon - error - rubrics unknown upto now were not applied - fixed
As the Antiphonale Monasticum contains only very spare rubrics (is intended to be used by people
familiar with rubrics contained in the monastic breviary, which I don't have),
I didn't know, that whenever the ant.s for vespers are taken from the lauds,
the 4th ant. is left out and the 5th ant. is to be sung with the 4th psalm.
Now I know.